Updated TableManager to not create a table refresh thread as it is most
likely unnecessary as we might not have enough tables to warrant a
separate thread. Chat was updated to display outgoing messages as soon
as they are sent instead of waiting for a receiving message in order to
make the client seem more "snappy". Code is a pretty big mess now
though.